dev-build/ninja: fix -jN vs --jobserver-pool interaction
Before this patch, we get:
```
$ ninja -v -j32 -l32 --jobserver-pool install
ninja: Creating jobserver pool for 32 parallel jobs
ninja: warning: ignoring jobserver: Explicit parallelism specified []
```

We don't want that. We want to be subervient to a pool if it exists,
spawn one if it doesn't, and we want to consume up to 32 tokens if
available (either in the jobserver we just spawned, or in a parent one).